Add "NodePosition" view for getting position after submitting form
