+++ navigatAR/navigatAR/ManageNodesViewController.swift
@@ -1,0 +1,26 @@
+//
+//  ManageNodesViewController.swift
+//  navigatAR
+//
+//  Created by Michael Gira on 2/3/18.
+//  Copyright © 2018 MICDS Programming. All rights reserved.
+//
+
+import UIKit
+
+class ManageNodesViewController: UIViewController {
+
+	@IBAction func unwindToManageNodes(segue: UIStoryboardSegue) { }
+
+	override func viewDidLoad() {
+		super.viewDidLoad()
+		// Do any additional setup after loading the view, typically from a nib.
+	}
+
+	override func didReceiveMemoryWarning() {
+		super.didReceiveMemoryWarning()
+		// Dispose of any resources that can be recreated.
+	}
+
+}
+
